Refactor: AdminViewController methods to use Load prefix
This commit refactors the methods in the AdminViewController to use a "Load" prefix for better clarity and consistency. This change improves the readability and maintainability of the code by clearly indicating that these methods are responsible for loading specific views or data.

changes also moving a method from `AdminViewController` a invokable  `GetPendingProjectController` class