Add signed middleware to document form routes and remove unused viewSR route
This commit adds the 'signed' middleware to the routes that handle form submissions for TNA, project proposal, and RTEC report documents to enhance security by ensuring requests are coming from authorized sources. It also removes an unused route ('/viewSR') that was pointing to a status report view.

changes also include:
- adding form attribute to the `Set TNA Form` button on the TNA